pyproject.toml
[project]
name = "mmr"
version = "0.1.0"
description = "Make Me Rich - Algorithmic Trading Platform"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.9",
    "backoff>=2.2",
    "beautifulsoup4>=4.12",
    "bs4",
    "click>=8.1",
    "click-help-colors>=0.9",
    "click-option-group>=0.5",
    "click-repl>=0.3",
    "cloup>=3.0",
    "coloredlogs>=15.0",
    "configobj>=5.0",
    "crontab>=1.0",
    "dateparser>=1.2",
    "dill>=0.3.7",
    "duckdb>=1.0",
    "durations>=0.3",
    "eventkit>=1.0.0",
    "exchange-calendars>=4.5",
    "expression>=5.0",
    "ib_async>=1.0.0",
    "ipython>=8.20",
    "lxml>=5.0",
    "lz4>=4.3",
    "matplotlib>=3.8",
    "msgpack>=1.0.7",
    "numpy>=1.26",
    "pandas>=2.1",
    "plotext>=5.2",
    "plotille>=5.0",
    "plotly>=5.18",
    "prompt-toolkit>=3.0",
    "psutil>=5.9",
    "pyarrow>=17.0",
    "python-dateutil>=2.8",
    "pyfiglet>=1.0",
    "pypager>=3.0",
    "pytz>=2024.1",
    "pyyaml>=6.0",
    "pyzmq>=26.0",
    "reactivex>=4.0.4",
    "requests>=2.31",
    "rich>=13.7",
    "scikit-learn>=1.4",
    "scipy>=1.12",
    "setproctitle>=1.3",
    "tabulate>=0.9",
    "textual>=0.40",
    "tornado>=6.4",
    "uniplot>=0.12",
    "vectorbt>=0.26",
    "yfinance>=0.2",
    "massive>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.8",
    "isort>=5.13",
]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
]

[project.scripts]
trader-service = "trader.trader_service:main"
strategy-service = "trader.strategy_service:main"
data-service = "trader.data_service:main"
mmr = "trader.mmr_cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["trader*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.isort]
multi_line_output = 3
force_alphabetical_sort = true
from_first = false
line_length = 110
lines_after_imports = 2
